+++ kubernetes/bootstrap.sh
@@ -3,6 +3,7 @@ set -x
 CLUSTER_NAME=$1
 ZONE=$2
 PROJECT=$3
+EXTERNAL_IP=$4
 SCRIPT_DIR=$(dirname $(realpath "$0"))
 BOOTSTRAP_DIR="${SCRIPT_DIR}"/bootstrap
 CHART_DIR="${SCRIPT_DIR}"/charts
@@ -13,5 +14,14 @@ gcloud beta container clusters update "${CLUSTER_NAME}" --zone "${ZONE}" --loggi
 
 kubectl apply -f "${BOOTSTRAP_DIR}"/tiller.yaml
 helm init --wait --service-account tiller
-helm upgrade --install nginx stable/nginx-ingress
+
 helm dependency update "${CHART_DIR}"/search_engine/
+
+helm repo add gitlab https://charts.gitlab.io/
+helm repo update
+helm upgrade --wait --install gitlab gitlab/gitlab \
+  --timeout 600 \
+  --values "${CHART_DIR}"/gitlab/values.yaml \
+  --set global.hosts.externalIP="${EXTERNAL_IP}"
+
+echo "Gitlab root password: $(kubectl get secret gitlab-gitlab-initial-root-password -ojsonpath={.data.password} | base64 --decode ; echo)"